Add files via upload
this is finished project for d20, in fact the roll amount of time doesn't make the output the same so you can use the same roll time number amount and it will output with a random (as random random number generators can be) every time! this is a really cool thing that was made and I am really happy it exists.